Add Logging

+++ src/main/java/team1/myshop/web/CategoryService.java
@@ -1,5 +1,6 @@
 package team1.myshop.web;
 
+import org.apache.logging.log4j.LogManager;
 import team1.myshop.contracts.UserRights;
 import team1.myshop.web.model.Category;
 import team1.myshop.web.helper.JsonParser;
@@ -11,14 +12,19 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import java.util.Collection;
 
-import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
-import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
+
+import static javax.servlet.http.HttpServletResponse.*;
 
 @Path("/categories")
 public class CategoryService extends ServiceBase {
 
     public CategoryService() {
         super();
+    }
+
+    @Override
+    public void initializeLogger() {
+        this.setLogger(LogManager.getLogger(CategoryService.class));
     }
 
     @GET
@@ -77,7 +83,10 @@ public class CategoryService extends ServiceBase {
         this.initialize();
 
         // check user rights
-        auth.ensureUserRight(request, response, UserRights.CAN_EDIT_CATEGORY);
+        if(!auth.userHasRight(request, UserRights.CAN_EDIT_CATEGORY)){
+            http.cancelRequest(response, SC_UNAUTHORIZED);
+            return null;
+        }
 
         Category cat = JsonParser.parse(catString, Category.class);
 
@@ -101,12 +110,20 @@ public class CategoryService extends ServiceBase {
                                @Context HttpServletResponse response) {
         this.initialize();
 
+        logger.debug("User tries to delete category "+category );
+
         // check user rights
-        auth.ensureUserRight(request, response, UserRights.CAN_DELETE_CATEGORY);
+        if (!auth.userHasRight(request, UserRights.CAN_DELETE_CATEGORY)) {
+            logger.info("User tried to delete category " + category + " but did not have the right" + category);
+            http.cancelRequest(response, SC_UNAUTHORIZED);
+            return;
+        }
 
         // delete category
+        logger.debug("User has the right to delete category " + category);
         dh.deleteCategory(category);
+        logger.debug("User has deleted category " + category);
 
-        response.setStatus(HttpServletResponse.SC_NO_CONTENT);
+        response.setStatus(SC_NO_CONTENT);
     }
 }
